feat: refactor split QdrantService into domain modules and extract utilities
- Split services/QdrantService.ts (502 lines) into:
  - services/CollectionService.ts - collection CRUD operations
  - services/PointsService.ts - points upsert/search/delete
  - services/errors.ts - QdrantServiceError class

- Extract utilities from repositories:
  - utils/normalization.ts - vector extraction, search body parsing
  - utils/distance.ts - KNN and index distance mapping
  - utils/retry.ts - exponential backoff for transient YDB errors

- Refactor package/api.ts with buildTenantClient factory (-60 lines)

- Update all imports in routes, tests, and documentation
